feat(计时器): 添加计时器蓝图模块并重构相关功能
将计时器相关功能从主应用拆分为独立蓝图模块，包含以下主要变更：
1. 创建独立蓝图模块
2. 添加README说明计时器工作流程
