+++ point_and_timer_blueprint/README.md
@@ -1,0 +1,18 @@
+graph TD
+    start[开始] --> A[来到/point路由]
+    A --> B{是否是定时任务（时间不为0）?}
+    
+    B -->|是| C[调用timer函数跳转计时器]
+    C --> D[计时完成，再次回到/point路由]
+    D --> E[完成相应积分变更（非重复任务此时删除）]
+    E --> F{任务是否被删除?}
+    
+    F -->|是| G[渲染正常积分操作结果页面，自动跳转回主页]
+    F -->|否| H[返回积分操作结果页面，传递from_page = timer]
+    H --> I[触发渲染包含询问是否继续的表单]
+    I --> J{用户选择是否继续?}
+    J -->|继续| K[发送post请求到/timer_submit路由，参数转到timer函数]
+    K --> C[开始计时]
+    
+    B -->|否| L[调用process_point_change函数处理积分变更]
+    L --> M[渲染正常积分操作结果页面，自动跳转回主页]
